Use multiboot module to compare in-memory images
The existing 'multiboot' module is hacked to compare the CSL generated
memory image with the one generated via the existing multiboot loader.
This is used to test image equivalence during CI.

Note: It is assumed that the memory relocator allocates chunks for both
images consecutively. The comparison might fail if uninitialized bytes
exist between memory regions (these will be zero for multiboot and
undefined for CSL).