grub-shell: Improve mkrescue_args handling
Use an array to store mkrescue_args. Required to handle mkrescue
arguments which expect a whitespace separated list as argument value,
e.g.:

--mkrescue-arg="--modules=normal serial pgp"

Without this patch the shell does not correctly pass on the arguments to
mkrescue, because the second module is interpreted as a separate
argument.

Quoting the modules does not help either as this affects the way the
shell passes on the arguments to mkrescue (i.e. the quote becomes part
of the first module name).