CSL: Add highmem (> 4 GiB) support
Implement support for > 32 bit destination addresses to
i386-pc/x86_64-efi CSL fill and write commands.

=== i386-pc

To use highmem addresses on i386-pc, PAE paging is required. Implement
support for it by porting the code from FILO bootloader and adjust it to
work with GRUB mm/relocator.

A relocator is used to acquire memory for the 2 MiB PAE window to
prevent overlap with regions allocated by GRUB (grub_malloc and
friends), and to fulfill the 2 MiB alignment constraint. The relocator
is freed after each fill|write in order to avoid occupying physical
memory which might be targeted by an actual fill|write cmd.

In contrast to memset_pae and the read_pae implementation in FILO, a
temporary buffer is required for grub_file_read because GRUB, depending
on the boot media, might switch to realmode to actually read the data.
As this breaks PAE, use a temporary buffer with PAE disabled and copy
the data to the destination in a second step.

The same realmode problem arises with VGA output (grub_printf): GRUB
uses realmode and grub_bios_interrupt in the VGA terminal/console code.
Therefore no log output must be produced during the time period when
PAE is active. The period is kept as small as possible. Another option
would be to use virtual-8086 mode, but that seems too invasive for now.

=== x86_64-efi

On a system with UEFI, paging mode is enabled by firmware and any memory
space defined by the UEFI memory map is identity mapped (see UEFI system
specification, section "2.3.4. x64 Platforms"). This allows direct
addressing of regions above 4 GiB, even without relocator. This works
because GRUB does not malloc() objects in > 4 GiB.

Identical to i386-pc, addresses below the 32 bit boundary must be
processed with a relocator in order to avoid overlap with GRUB mm
objects.